+++ packages/dumb/src/components/stats/average-daily-dose/average-daily-dose-stat.tsx
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022, Diabeloop
+ * Copyright (c) 2022-2023, Diabeloop
  *
  * All rights reserved.
  *
@@ -32,8 +32,11 @@ import { formatDecimalNumber } from '../../../utils/format/format.util'
 import { useTranslation } from 'react-i18next'
 import { SimpleValue } from '../common/simple-value'
 import { EMPTY_DATA_PLACEHOLDER } from '../../../models/stats.model'
+import { StatTooltip } from '../../tooltips/stat-tooltip/stat-tooltip'
+import Box from '@mui/material/Box'
 
 interface AverageDailyDoseStatProps {
+  annotations: string[]
   dailyDose: number
   footerLabel: string
   title: string
@@ -43,6 +46,7 @@ interface AverageDailyDoseStatProps {
 
 const AverageDailyDoseStat: FunctionComponent<AverageDailyDoseStatProps> = (props) => {
   const {
+    annotations,
     dailyDose,
     footerLabel,
     title,
@@ -66,9 +70,10 @@ const AverageDailyDoseStat: FunctionComponent<AverageDailyDoseStatProps> = (prop
     <div className={commonStyles.statWrapper}>
       <div className={`${commonStyles.stat} ${styles.isOpen}`}>
         <div className={commonStyles.statHeader}>
-          <div className={commonStyles.chartTitle}>
+          <Box display="flex" className={commonStyles.chartTitle}>
             {title}
-          </div>
+            <StatTooltip annotations={annotations} />
+          </Box>
           <SimpleValue
             className={styles.insulinTitle}
             suffix={t('U')}