+++ packages/dumb/src/components/stats/cbg-percentage/cbg-percentage-bar.util.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022, Diabeloop
+ * Copyright (c) 2023, Diabeloop
  *
  * All rights reserved.
  *
@@ -25,31 +25,19 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-.legend-title {
-  font-size: 12px;
+import { UnitsType } from '../../../models/enums/units-type.enum'
+
+export const ensureNumeric = (value: number | undefined | null): number => {
+  return !value || isNaN(value) ? 0 : value
 }
 
-.stat-footer {
-  display: flex;
-  flex-wrap: nowrap;
-  align-items: center;
-  justify-content: space-between;
-  border-top: 1px solid var(--stat--border);
-  padding: 0.25em 0.625em .375em;
-  overflow: hidden;
-}
-
-.title {
-  margin-left: 4px;
-  height: 21px;
-}
-
-.tooltip-icon {
-  position: relative;
-  top: -0.3em;
-  margin-left: 0.1em;
-}
-
-.tooltip-icon > img {
-  width: 0.9em;
+export const formatBgValue = (value: number, units: UnitsType): string => {
+  const valueToString = Math.round(value).toString()
+  switch (units) {
+    case UnitsType.MMOLL:
+      return parseFloat(valueToString).toFixed(1)
+    case UnitsType.MGDL:
+    default:
+      return parseInt(valueToString).toString()
+  }
 }