+++ packages/yourloops/components/alarm/alarms-content-configuration.hook.tsx
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022, Diabeloop
+ * Copyright (c) 2022-2023, Diabeloop
  *
  * All rights reserved.
  *
@@ -32,9 +32,9 @@ import PatientUtils from '../../lib/patient/patient.util'
 import { type Patient } from '../../lib/patient/models/patient.model'
 import { useTranslation } from 'react-i18next'
 import { buildThresholds, isInvalidPercentage, REGEX_VALUE_BG } from './alarm-content-configuration.utils'
-import { UnitsType } from '../../lib/units/models/enums/units-type.enum'
+import { UnitsType } from 'dumb'
 import { type Thresholds } from '../../lib/patient/models/alarms.model'
-import { DEFAULT_BG_VALUES } from '../alarm/alarms.default'
+import { DEFAULT_BG_VALUES } from './alarms.default'
 
 export interface AlarmsContentConfigurationHookProps {
   monitoring: Monitoring