+++ packages/yourloops/test/unit/lib/auth/auth.hook.test.tsx
@@ -41,7 +41,7 @@ import { AuthenticatedUserMetadata } from '../../../../lib/auth/models/enums/aut
 import { UserRoles } from '../../../../lib/auth/models/enums/user-roles.enum'
 import { type UserMetadata } from '../../../../lib/auth/models/user-metadata.model'
 import { CountryCodes } from '../../../../lib/auth/models/country.model'
-import { UnitsType } from '../../../../lib/units/models/enums/units-type.enum'
+import { UnitsType } from 'dumb'
 import { LanguageCodes } from '../../../../lib/auth/models/enums/language-codes.enum'
 
 jest.mock('@auth0/auth0-react')